Abastract
Create an abstract class called Animal with the following properties:

    name: String
    age: int
    species: String

The Animal class should also have the following abstract methods:

    makeSound()
    eat()
    sleep()

Create two subclasses of Animal: Dog and Cat. The Dog class should override the makeSound() method to return the string "Woof!" and the Cat class should override the makeSound() method to return the string "Meow!".
The Dog class should also override the eat() method to return the string "I eat dog food!" and the Cat class should override the eat() method to return the string "I eat cat food!".
The Dog class should also override the sleep() method to return the string "I sleep on the couch!" and the Cat class should override the sleep() method to return the string "I sleep in a box!".
Create an instance of each subclass and call the makeSound(), eat(), and sleep() methods on each instance.